requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
pyaml==25.7.0
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.2
sentry-sdk==2.42.1
setuptools==80.9.0
smmap==5.0.2
stim==1.15.0
sympy==1.14.0
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-cluster==1.6.3
torch-spline-conv==1.2.2
torch-geometric==2.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wandb==0.22.2
xxhash==3.6.0
yarl==1.22.0
